feat: rename folder and remaining vars
