+++ .env.test
@@ -16,9 +16,9 @@ CURRICULUM_API_2023_URL="http://api.example.com"
 CURRICULUM_API_AUTH_TYPE="auth_type"
 CURRICULUM_API_AUTH_KEY="very secret"
 CURRICULUM_API_2023_AUTH_KEY="very secret"
-PERSONALISATION_API_URL="test-url"
-PERSONALISATION_API_AUTH_KEY="key"
-PERSONALISATION_API_AUTH_ROLE="role"
+EDUCATOR_API_URL="test-url"
+EDUCATOR_API_AUTH_KEY="key"
+EDUCATOR_API_AUTH_ROLE="role"
 CLERK_SIGNING_SECRET="secret"
