+++ .eslintignore
@@ -8,6 +8,6 @@ out
 storybook-static
 src/node-lib/curriculum-api-2023/generated
 src/node-lib/sanity-graphql/generated
-src/node-lib/personalisation-api/generated
+src/node-lib/educator-api/generated
 
 !.github
