+++ package.json
@@ -24,7 +24,7 @@
     "format:check": "prettier '**/*.{j,t}{s,sx}' --check",
     "gql-codegen:curriculum-2023": "graphql-codegen --config src/node-lib/curriculum-api-2023/codegen.yml -r dotenv/config",
     "gql-codegen:sanity": "graphql-codegen --config src/node-lib/sanity-graphql/codegen.yml -r dotenv/config",
-    "gql-codegen:personalisation": "graphql-codegen --config src/node-lib/personalisation-api/codegen.yml -r dotenv/config",
+    "gql-codegen:educator": "graphql-codegen --config src/node-lib/educator-api/codegen.yml -r dotenv/config",
     "image-data:all": "./scripts/build/image_data/all.sh",
     "lint": "NODE_ENV=test next lint",
     "lint:precommit": "eslint --fix",
