+++ src/app/api/webhooks/route.ts
@@ -4,7 +4,7 @@ import { WebhookEvent } from "@clerk/nextjs/server";
 import { NextRequest } from "next/server";
 
 import getServerConfig from "@/node-lib/getServerConfig";
-import { getWebhookEducatorApi } from "@/node-lib/personalisation-api";
+import { getWebhookEducatorApi } from "@/node-lib/educator-api";
 import errorReporter from "@/common-lib/error-reporter";
 
 export async function POST(req: NextRequest) {
