+++ src/node-lib/getServerConfig.ts
@@ -151,22 +151,22 @@ const envVars = satisfies<Record<string, EnvVar>>()({
     default: null,
   },
   educatorApiUrl: {
-    value: process.env.PERSONALISATION_API_URL,
-    envName: "PERSONALISATION_API_URL",
+    value: process.env.EDUCATOR_API_URL,
+    envName: "EDUCATOR_API_URL",
     required: true,
     availableInBrowser: false,
     default: null,
   },
   educatorApiAuthKey: {
-    value: process.env.PERSONALISATION_API_AUTH_KEY,
-    envName: "PERSONALISATION_API_AUTH_KEY",
+    value: process.env.EDUCATOR_API_AUTH_KEY,
+    envName: "EDUCATOR_API_AUTH_KEY",
     required: true,
     availableInBrowser: false,
     default: null,
   },
   educatorApiAuthRole: {
-    value: process.env.PERSONALISATION_API_AUTH_ROLE,
-    envName: "PERSONALISATION_API_AUTH_ROLE",
+    value: process.env.EDUCATOR_API_AUTH_ROLE,
+    envName: "EDUCATOR_API_AUTH_ROLE",
     required: true,
     availableInBrowser: false,
     default: null,
